+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChessTournament-Bundle Changelog
 
+## Version 0.0.2 (2023-11-30)
+
+* Add: Abhängigkeit codefog/contao-haste
+* Change: Toggle-Funktion durch Haste-Toggler ersetzt
+* Add: published-Felder in tl_chesstournament_players und tl_chesstournament_results
+
 ## Version 0.0.1 (2021-11-19)
 
 * Migration der Contao-3-Erweiterung